+++ pyccel/ast/builtin_methods/list_methods.py
@@ -14,9 +14,13 @@ from pyccel.ast.datatypes import NativeVoid, NativeGeneric, NativeHomogeneousLis
 from pyccel.ast.internals import PyccelInternalFunction
 
 
-__all__ = ('ListAppend','ListPop')
+__all__ = ('ListAppend',
+           'ListClear',
+           'ListInsert',
+           'ListPop',
+           )
 
-
+#==============================================================================
 class ListAppend(PyccelInternalFunction):
     """
     Represents a call to the .append() method.
@@ -35,7 +39,7 @@ class ListAppend(PyccelInternalFunction):
     list_variable : Variable
         The variable representing the list.
     
-    new_elem : Variable
+    new_elem : TypedAstNode
         The argument passed to append() method.
     """
     __slots__ = ("_list_variable", "_append_arg")
@@ -45,7 +49,7 @@ class ListAppend(PyccelInternalFunction):
     _order = None
     _rank = 0
     _precision = -1
-    _class_type = NativeHomogeneousList()
+    _class_type = NativeVoid()
     name = 'append'
 
     def __init__(self, list_variable, new_elem) -> None:
@@ -130,3 +134,117 @@ class ListPop(PyccelInternalFunction) :
         """
         return self._list_variable
 
+#==============================================================================
+class ListClear(PyccelInternalFunction) :
+    """
+    Represents a call to the .clear() method.
+    
+    Represents a call to the .clear() method which deletes all elements from a list, 
+    effectively turning it into an empty list.
+    Note that the .clear() method doesn't return any value.
+
+    Parameters
+    ----------
+    list_variable : TypedAstNode
+        The name of the list.
+    """
+    __slots__ = ('_list_variable',)
+    _attribute_nodes = ('_list_variable',)
+    _dtype = NativeVoid()
+    _precision = -1
+    _rank = 0
+    _order = None
+    _shape = None
+    _class_type = NativeVoid()
+    name = 'clear'
+
+    def __init__(self, list_variable):
+        self._list_variable = list_variable
+        super().__init__()
+
+    @property
+    def list_variable(self):
+        """
+        Provide the name of the list as the return value.
+
+        Provide the name of the list as the return value.
+        """
+        return self._list_variable
+
+#==============================================================================
+class ListInsert(PyccelInternalFunction):
+    """
+    Represents a call to the .insert() method.
+
+    Represents a call to the .insert() method of an object with a list type,
+    which inserts a given element at a given index in a list.
+    This method returns `None`.
+    The insert method is called as follows:
+
+    >>> a = [2, 3, 4]
+    >>> a.insert(0, 1)
+    >>> print(a)
+    [1, 2, 3, 4]
+
+    Parameters
+    ----------
+    list_variable : Variable
+        The variable representing the list.
+
+    index : TypedAstNode
+        The index value for the element to be added.
+    
+    new_elem : TypedAstNode
+        The argument passed to insert() method.
+    """
+    __slots__ = ("_index", "_list_variable", "_insert_arg")
+    _attribute_nodes = ("_index", "_list_variable", "_insert_arg")
+    _dtype = NativeVoid()
+    _shape = None
+    _order = None
+    _rank = 0
+    _precision = -1
+    _class_type = NativeVoid()
+    name = 'insert'
+
+    def __init__(self, list_variable, index, new_elem) -> None:
+        is_homogeneous = (
+            new_elem.dtype is not NativeGeneric() and
+            list_variable.dtype is not NativeGeneric() and
+            list_variable.dtype == new_elem.dtype and
+            list_variable.precision == new_elem.precision and
+            list_variable.rank - 1 == new_elem.rank
+        )
+        if not is_homogeneous:
+            raise TypeError("Expecting an argument of the same type as the elements of the list")
+        self._index = index
+        self._list_variable = list_variable
+        self._insert_arg = new_elem
+        super().__init__()
+
+    @property
+    def index(self):
+        """
+        Index in which the element will be added.
+
+        Index in which the element will be added.
+        """
+        return self._index
+
+    @property
+    def list_variable(self):
+        """
+        Get the variable representing the list.
+
+        Get the variable representing the list.
+        """
+        return self._list_variable
+
+    @property
+    def insert_argument(self):
+        """
+        Get the argument which is passed to insert().
+
+        Get the argument which is passed to insert().
+        """
+        return self._insert_arg