+++ CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
--   #1529 : Add Python support for list method `append()`
--   #1690 : Add support for `pop()` method of list class
+-   #1689 : Add Python support for list method `append()`
+-   #1692 : Add Python support for list method `insert()`
+-   #1690 : Add Python support for list method `pop()`
+-   #1691 : Add Python support for list method `clear()`
 -   #1575 : Add support for homogeneous tuple type annotations on variables.
+-   #1425 : Add support for `numpy.isnan`, `numpy.isinf` and `numpy.isfinite`.
 
 ### Fixed